@@ -1,5 +1,4 @@
-import _Vue = require("vue");
-import { WatchOptions } from "vue";
+import _Vue, { WatchOptions } from "vue";
 
 // augment typings of Vue.js
 import "./vue";
@@ -20,8 +19,8 @@ export declare class Store<S> {
   subscribe<P extends MutationPayload>(fn: (mutation: P, state: S) => any): () => void;
   watch<T>(getter: (state: S) => T, cb: (value: T, oldValue: T) => void, options?: WatchOptions): () => void;
 
-  registerModule<T>(path: string, module: Module<T, S>): void;
-  registerModule<T>(path: string[], module: Module<T, S>): void;
+  registerModule<T>(path: string, module: Module<T, S>, options?: ModuleOptions): void;
+  registerModule<T>(path: string[], module: Module<T, S>, options?: ModuleOptions): void;
 
   unregisterModule(path: string): void;
   unregisterModule(path: string[]): void;
@@ -82,8 +81,14 @@ export interface StoreOptions<S> {
   strict?: boolean;
 }
 
+type ActionHandler<S, R> = (injectee: ActionContext<S, R>, payload: any) => any;
+interface ActionObject<S, R> {
+  root?: boolean;
+  handler: ActionHandler<S, R>;
+}
+
 export type Getter<S, R> = (state: S, getters: any, rootState: R, rootGetters: any) => any;
-export type Action<S, R> = (injectee: ActionContext<S, R>, payload: any) => any;
+export type Action<S, R> = ActionHandler<S, R> | ActionObject<S, R>;
 export type Mutation<S> = (state: S, payload: any) => any;
 export type Plugin<S> = (store: Store<S>) => any;
 
@@ -94,6 +99,10 @@ export interface Module<S, R> {
   actions?: ActionTree<S, R>;
   mutations?: MutationTree<S>;
   modules?: ModuleTree<R>;
+}
+
+export interface ModuleOptions{
+  preserveState?: boolean
 }
 
 export interface GetterTree<S, R> {
@@ -111,3 +120,9 @@ export interface MutationTree<S> {
 export interface ModuleTree<R> {
   [key: string]: Module<any, R>;
 }
+
+declare const _default: {
+  Store: typeof Store;
+  install: typeof install;
+}
+export default _default;
